fix bug

+++ prompt/plot.py
@@ -24,7 +24,7 @@ def get_prompt_plot(task_id ,method):
             match = re.search(r"评估分数：([\d.]+)/10",evals[i].output_text)
             if match:
                 x_values.append(i+1)
-                y_values.append(int(match.group(1)))
+                y_values.append(float(match.group(1)))
 
         # 设置中文字体（根据系统选择）
         plt.rcParams['font.sans-serif'] = ['SimHei']  # Windows
